Add the Tock Workshop initial track (#1)
* add the Tock Workshop inital track

* added prerequisites

* fixed introduction typo

* architecture WIP

* added build capsule task

* Small fixes

This commit adds the prerequisites for `probe-rs` and adds some
clarifications regarding some of the necessary actions.

* fix: added list output

* added test application task

* nit: fixed format on list

* added periodic print task

* added thermistor task

* Add GitHub links for the repositories

* Small fixes

This commit adds small fixes based on the received feedback.

* Add NixOS alternative for TockWorld.

This commit adds a NixOS VM alternative for Windows users or
user that don't want to follow the prerequisites. The text of
the workshop has been modified to reflect this addition.

---------

Co-authored-by: George-Dănuț Aldea <danutz.aldea23@gmail.com>
Co-authored-by: Darius-Andrei Jipa <darius.jipa@oxidos.io>
Co-authored-by: Darius-Andrei Jipa <jipad16@yahoo.com>
Co-authored-by: Darius-Andrei Jipa <53440061+JADarius@users.noreply.github.com>